+++ CHANGELOG.md
@@ -8,6 +8,9 @@ This project <em>does not yet</em> adheres to [Semantic Versioning](https://semv
 ## [unreleased]
 
 ### Added
+- New interactive key task developed in the interactive_key branch [#1810] 
+- Added new model for matrix based interactive keys which produce JSON 
+    for the interactive key task [#1810]
 - Handling for new unicode minutes, seconds symbols [#1526]
 - Descriptor object radial paths
 - Many specs related to dynamic observation matrix items